requirements.txt
Django==1.8.5
django-wpadmin==1.7.4
djangorestframework==3.2.4
psycopg2==2.6.1
wheel==0.24.0
